Clean up logic for padding character state vectors (remove branches).
Use __attribute__ ((aligned (16))) in order to be able to use the movdqa
instruction rather than the movdqu instruction.

Fix a bug in the loop variable initialization for tr_p_mp_ia32_[fp]score().

Move score updating inside the inner loop of tr_p_mp_ia32_fscore(), in order to
be able to compare the current score against the maximum interesting score.

Terminate *_fscore() if a maximum score is exceeded.

Do not calculate scores for any trees based an a reconnection of a subtree that
exceeds the maximum interesting score.