Add mixture model support to Mc3.
Add the rmult parameter to Lik, so that each Q matrix within a mixture can be
weighted as a whole.

Remove Rcov plotting.

Do not print old Rcov values.

Replace Mc3's cvgDelay with cvgSampStride.

Allow there to be a single run (and disable convergence diagnostics in that
case).

Fix weighted lnL computation.